predator稳定测试版

+++ cheetah-predator/src/main/java/cheetah/predator/core/SessionRegistry.java
@@ -30,5 +30,5 @@ public interface SessionRegistry {
      */
     void unregister(Session session);
 
-    long size();
+    int size();
 }
